requirements.txt
asgiref==3.4.1
Django==3.2.6
django-hosts==4.0
djangorestframework==3.12.4
gunicorn==20.1.0
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
cryptography==2.8
django-extensions
djangorestframework-simplejwt
django-allauth
